Decompose mobile app presentation pieces
